# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anime Gallery Media Server",
    "description": "\nAnime Gallery Media Server API 覆盖前台图片浏览、后台媒体资料维护、批量上传任务、\n元数据导入、部署设置与系统健康检查。\n\n鉴权模型：\n\n- 浏览器后台会话使用 `agms_admin_session` HttpOnly Cookie。\n- 运维脚本建议使用 `Authorization: Bearer <AGMS_API_KEYS value>`，并只授予任务实际需要的权限。\n- 使用 Cookie 会话调用非安全方法时，还需要提交与 `agms_admin_csrf` Cookie 匹配的 `X-CSRF-Token`。\n- API Key 不依赖 Cookie，因此不需要 CSRF；可在后台独立创建、设置到期时间、调整权限或立即撤销。\n- 旧版本升级后，已有 API Key 默认迁移为全部权限，避免现有自动化任务中断。\n",
    "version": "0.1.5"
  },
  "servers": [
    {
      "url": "/",
      "description": "当前部署域名"
    },
    {
      "url": "/api",
      "description": "API 前缀"
    }
  ],
  "paths": {
...
    "/api/images/upload": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "立即上传一张或多张图片",
        "operationId": "upload_images_api_images_upload_post",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_images_api_images_upload_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "上传文件，校验后缀与内容，支持静态 SDR 原图转 WebP、动图保留原格式、HDR JXR 转 HDR AVIF。\n\n所需 API Key 权限：`uploads:manage`。",
        "x-api-key-scopes": [
          "uploads:manage"
        ]
      }
    },
...
    "/api/upload-tasks": {
      "post": {
        "tags": [
          "upload-tasks"
        ],
        "summary": "创建批量上传任务",
        "operationId": "create_upload_tasks_api_upload_tasks_post",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_upload_tasks_api_upload_tasks_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTaskCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "先校验整批文件、关联 ID 和 SHA-256，再以单个事务加入服务端并发处理队列。任一文件校验或任务提交失败时，本批次不会留下任务或暂存文件。\n\n所需 API Key 权限：`uploads:manage`。",
        "x-api-key-scopes": [
          "uploads:manage"
        ]
      },
      "get": {
        "tags": [
          "upload-tasks"
        ],
        "summary": "获取上传任务列表",
        "operationId": "list_upload_tasks_api_upload_tasks_get",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ids"
            },
            "description": "可选，逗号分隔的上传任务 ID 列表，用于直接查询指定任务。"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            },
            "description": "从 1 开始的页码。"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            },
            "description": "每页记录数。大多数列表接口最多返回 100 条。"
          },
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTaskListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "分页返回上传任务，可按排队、运行、等待重试、成功、失败或已取消状态筛选。\n\n所需 API Key 权限：`uploads:manage`。",
        "x-api-key-scopes": [
          "uploads:manage"
        ]
      }
    },
    "/api/upload-tasks/check-duplicates": {
      "post": {
        "tags": [
          "upload-tasks"
        ],
        "summary": "检查重复哈希",
        "operationId": "check_upload_duplicates_api_upload_tasks_check_duplicates_post",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadDuplicateCheckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadDuplicateCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "批量比对客户端计算的 SHA-256，分别标记图库已有、上传队列已有和本批次内重复。服务端仍会在最终写入时执行唯一性保护。\n\n所需 API Key 权限：`uploads:manage`。",
        "x-api-key-scopes": [
          "uploads:manage"
        ]
      }
    },
    "/api/upload-tasks/check-duplicates-files": {
      "post": {
        "tags": [
          "upload-tasks"
        ],
        "summary": "检查重复上传文件",
        "operationId": "check_upload_duplicate_files_api_upload_tasks_check_duplicates_files_post",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_check_upload_duplicate_files_api_upload_tasks_check_duplicates_files_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadDuplicateCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "当客户端未计算哈希时，由服务端批量计算 SHA-256 并执行图库、队列和批次内重复预检。\n\n所需 API Key 权限：`uploads:manage`。",
        "x-api-key-scopes": [
          "uploads:manage"
        ]
      }
    },
    "/api/upload-tasks/{task_id}/retry": {
      "post": {
        "tags": [
          "upload-tasks"
        ],
        "summary": "重试上传任务",
        "operationId": "retry_task_api_upload_tasks__task_id__retry_post",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            },
            "description": "任务 ID。"
          },
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "立即重试仍保留暂存文件的失败任务，或提前执行等待重试的任务。\n\n所需 API Key 权限：`uploads:manage`。",
        "x-api-key-scopes": [
          "uploads:manage"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_images_api_images_upload_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          },
          "work_ids": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Ids"
          },
          "character_ids": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Character Ids"
          },
          "rating": {
            "type": "string",
            "title": "Rating",
            "default": "safe"
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public",
            "default": true
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "artist_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist Name"
          },
          "merge_duplicate_relations": {
            "type": "boolean",
            "title": "Merge Duplicate Relations",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_images_api_images_upload_post"
      },
...
      "ImageUploadResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ImageUploadResult"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ImageUploadResponse"
      },
...
    },
    "securitySchemes": {
      "OperationsApiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "AGMS_API_KEY",
        "description": "后台创建的可分权运维 API Key。权限：`library:read` 媒体库读取；`uploads:manage` 图片上传；`library:write` 资料维护；`library:delete` 永久删除；`system:read` 系统只读；`settings:manage` 系统设置；`updates:read` 更新只读；`updates:run` 执行更新；`api_keys:manage` API Key 管理"
      },
      "AdminSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agms_admin_session",
        "description": "浏览器管理员会话 Cookie。"
      },
      "CsrfTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agms_admin_csrf",
        "description": "管理员会话同时签发的 CSRF Cookie。"
      },
      "CsrfTokenHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "使用 Cookie 会话调用非安全方法时，必须与 CSRF Cookie 一致。"
      }
    }
  },
...
}